+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Change commands that run the developer tools to require the code to be loaded first
 * Inform the user when commands that spawn processes complete
 * Add experimental code lens support for test results
+* Add experimental code lens support for entity cyclomatic complexity
+* Add "Logtalk: Compute Metrics" command
 * Add "Logtalk: Toggle Code Lens" command
 * Add "Logtalk: Generate Documentation (workspace)" command
 * Add "Logtalk: Generate Diagrams (workspace)" command